healthcare_mobile/requirements.txt
kivy>=2.2.0
kivymd>=1.1.1
requests>=2.28.0
cryptography>=40.0.0
qrcode>=7.4.0
pillow>=9.0.0
pyjwt>=2.6.0
keyring>=23.0.0
plyer>=2.0.0
